requirements.txt
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.24.0
pillow>=9.0.0
opencv-python>=4.7.0
scipy>=1.10.0
tqdm>=4.65.0
requests>=2.28.0
aiohttp>=3.8.0
python-dotenv>=1.0.0
pyyaml>=6.0
einops>=0.6.1
transformers>=4.30.0
accelerate>=0.20.0
safetensors>=0.3.1
torchsde>=0.2.6
spandrel>=0.1.0
kornia>=0.7.0
av>=10.0.0
pydantic>=2.0.0 